pyproject.toml
[tool.poetry]
name = "rest-class"
version = "0.1.0"
description = ""
authors = ["eqfwcev123 <dohyeonee95@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.11.0"
notebook = "^6.0.3"
django-extensions = "^2.2.8"
pygments = "^2.5.2"
django = "<3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
